+++ sh.py
@@ -24,7 +24,7 @@ http://amoffat.github.io/sh/
 #===============================================================================
 
 
-__version__ = "1.12.10"
+__version__ = "1.12.11"
 __project_url__ = "https://github.com/amoffat/sh"
 
 
@@ -398,7 +398,7 @@ class TimeoutException(Exception):
         self.exit_code = exit_code
         super(Exception, self).__init__()
 
-SIGNALS_THAT_SHOULD_THROW_EXCEPTION = (
+SIGNALS_THAT_SHOULD_THROW_EXCEPTION = set((
     signal.SIGABRT,
     signal.SIGBUS,
     signal.SIGFPE,
@@ -410,7 +410,7 @@ SIGNALS_THAT_SHOULD_THROW_EXCEPTION = (
     signal.SIGSEGV,
     signal.SIGTERM,
     signal.SIGSYS,
-)
+))
 
 
 # we subclass AttributeError because:
@@ -536,6 +536,7 @@ def which(program, paths=None):
     # if there's a path component, then we've specified a path to the program,
     # and we should just test if that program is executable.  if it is, return
     if fpath:
+        program = os.path.abspath(os.path.expanduser(program))
         if is_exe(program):
             found_path = program
 
@@ -805,10 +806,12 @@ class RunningCommand(object):
         """ here we determine if we had an exception, or an error code that we
         weren't expecting to see.  if we did, we create and raise an exception
         """
-        exc_class = get_exc_exit_code_would_raise(code, self.call_args["ok_code"])
+        ca = self.call_args
+        exc_class = get_exc_exit_code_would_raise(code, ca["ok_code"],
+                ca["piped"])
         if exc_class:
             exc = exc_class(self.ran, self.process.stdout, self.process.stderr,
-                    self.call_args["truncate_exc"])
+                    ca["truncate_exc"])
             raise exc
 
 
@@ -1399,7 +1402,8 @@ output"),
                 launch = lambda: os.spawnve(os.P_WAIT, cmd[0], cmd, call_args["env"])
 
             exit_code = launch()
-            exc_class = get_exc_exit_code_would_raise(exit_code, call_args["ok_code"])
+            exc_class = get_exc_exit_code_would_raise(exit_code,
+                    call_args["ok_code"], call_args["piped"])
             if exc_class:
                 if IS_PY3:
                     ran = " ".join([arg.decode(DEFAULT_ENCODING, "ignore") for arg in cmd])
@@ -1622,10 +1626,17 @@ def construct_streamreader_callback(process, handler):
     return fn
 
 
-def get_exc_exit_code_would_raise(exit_code, ok_codes):
+def get_exc_exit_code_would_raise(exit_code, ok_codes, sigpipe_ok):
     exc = None
     success = exit_code in ok_codes
     bad_sig = -exit_code in SIGNALS_THAT_SHOULD_THROW_EXCEPTION
+
+    # if this is a piped command, SIGPIPE must be ignored by us and not raise an
+    # exception, since it's perfectly normal for the consumer of a process's
+    # pipe to terminate early
+    if sigpipe_ok and -exit_code == signal.SIGPIPE:
+        bad_sig = False
+        success = True
 
     if not success or bad_sig:
         exc = get_rc_exc(exit_code)
@@ -1840,7 +1851,6 @@ class OProc(object):
         if IS_OSX:
             close_pipe_read, close_pipe_write = os.pipe()
 
-
         # session id, group id, process id
         self.sid = None
         self.pgid = None
@@ -1858,6 +1868,13 @@ class OProc(object):
                 # exits.  only ignore if we're backgrounded
                 if ca["bg"] is True:
                     signal.signal(signal.SIGHUP, signal.SIG_IGN)
+
+                # python ignores SIGPIPE by default.  we must make sure to put
+                # this behavior back to the default for spawned processes,
+                # otherwise SIGPIPE won't kill piped processes, which is what we
+                # need, so that we can check the error code of the killed
+                # process to see that SIGPIPE killed it
+                signal.signal(signal.SIGPIPE, signal.SIG_DFL)
 
                 # put our forked process in a new session?  this will relinquish
                 # any control of our inherited CTTY and also make our parent
@@ -2709,7 +2726,7 @@ class StreamWriter(object):
                 # sys.stdout.write(sys.stdin.read())
                 #
                 # then type 'a' followed by ctrl-d 3 times.  in python
-                # 2.6,2.7,3.3,3.4,3.5, it only takes 2 ctrl-d to terminate.
+                # 2.6,2.7,3.3,3.4,3.5,3.6, it only takes 2 ctrl-d to terminate.
                 # however, in python 3.1 and 3.2, it takes all 3.
                 #
                 # so here we send an extra EOF along, just in case.  i don't
@@ -3500,7 +3517,7 @@ if __name__ == "__main__": # pragma: no cover
 
         # if we're testing locally, run all versions of python on the system
         if action == "test":
-            all_versions = ("2.6", "2.7", "3.1", "3.2", "3.3", "3.4", "3.5")
+            all_versions = ("2.6", "2.7", "3.1", "3.2", "3.3", "3.4", "3.5", "3.6")
 
         # if we're testing on travis, just use the system's default python,
         # since travis will spawn a vm per python version in our .travis.yml